Matrix vector cleanup (#11)
* Use sparse matrix API to make sparse matrix access cleaner

* Use Vector/Matrix type aliases for arrays to be explicit regarding 1-dim and 2-dim structure

+++ test/ml100k.jl
@@ -11,7 +11,7 @@ info("Test dim: $(size(ml100kTest))")
 
 info("Test ML-100K - Training model...")
 fmMl100k = @time train(ml100kTrain, ml100kTrainRatings, 
-        method = Methods.sgd(num_epochs = UInt(10), alpha = 0.1),
+        method = Methods.sgd(num_epochs = 10, alpha = 0.1),
         model_params = Models.gauss(num_factors = 4))
 #info("FM Model: $fmMl100k")
 info("Test ML-100K - Prediction over test dataset...")
